// src/app/_core/models/Client.ts
export interface Client {
    id: number;
    // name: string;
    // phoneNumber: string;
    // password: string;
    // passwordConfirmation: string;
    // rolesId: number[];
    // idAgence: number;
    // firstname: string;
    // lastname: string;
    username: string;
    // phone: string;
    email: string;
    password: string;
    balance: number;
    enabled:boolean;
    // notificationToken: string;
}